separate meta proto
