+++ proto/meta.proto
@@ -1,0 +1,21 @@
+syntax = "proto3";
+
+package io.linkerd.proxy.meta;
+
+option go_package = "github.com/linkerd/linkerd2-proxy-api/go/meta";
+
+// General metadata about a configuration object. Typically references either an
+// implicit default configuration or a Kubernetes resource.
+message Metadata {
+  oneof kind {
+    string default = 1;
+    Resource resource = 2;
+  }
+
+  message Resource {
+    string group = 1;
+    string kind = 2;
+    string name = 3;
+  }
+}
+
